requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
load_dotenv
scikit-learn
scipy
pyaml
dill
mlflow
#-e .